requirements.txt
# requirements.txt

pandas==2.0.3
apache-airflow==2.8.2
pymongo==4.6.2
beautifulsoup4==4.12.3
requests==2.31.0
